Seeding doesnt work yet

+++ StudyMate/database/migrations/2020_03_17_210643_create_modules_table.php
@@ -16,13 +16,13 @@ class CreateModulesTable extends Migration
         Schema::create('modules', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->int('overseer');
-            $table->int('taught_by');
-            $table->int('followed_by');
-            $table->int('block_id');
-            $table->int('period_id');
+            $table->unsignedBigInteger('overseer');
+            $table->unsignedBigInteger('taught_by');
+            $table->unsignedBigInteger('followed_by');
+            $table->unsignedBigInteger('block_id');
+            $table->unsignedBigInteger('period_id');
             $table->date('year');
-            $table->int('study_points');
+            $table->integer('study_points');
             $table->boolean('is_finished');
         });
     }
